pyproject.toml
[project]
name = "informatique"
version = "0.1.0"
description = "Outils pédagogiques CM1-CM2 École Saint-Anne"
authors = [
    {name = "Johnny Xmas"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "odfpy (>=1.4.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
